# Conflicts:
#	pom.xml

+++ src/main/java/com/baichong/dao/mapper/LabelMapper.java
@@ -7,13 +7,16 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 /**
+ * 保证标签领域有统一收口
+ * 其他领域service曾统一调用com.baichong.service.LabelService
+ *
  * @author zhaoyongzhen
  * @since 2020/10/22 16:48
  */
 @Repository
 public interface LabelMapper {
 
-    int insert(LabelDO labelDO);
+    Long insert(LabelDO labelDO);
 
     LabelDO selectById(@Param("id") Long id);
